Change player move logics; Fix the Crystal and ball collider bug
Change player move logic - For the "a" and "w" moving methods, changed to make the player move parallel to the left or right instead of turning the player.
Fix the Crystal and ball collider bug - switched tiny box collider to fitted size sphere collider. Avoid hit and no vanish bug.